Discard cached plans when some UTILITY commands are run.
Intercept the UTILITY command being run and discard cached plan as needed.

There is no guarantee that the transaction will eventually commit, so we may
discard a plan for nothing, but this should hopfully not be a common case.
Also the current backend can't cache new plans until its transaction is
finished, as it could otherwise leave invalid or inefficient plans if its
transaction is rollbacked.  This is done by introducing a new
pg_shared_plans.read_only GUC, which is set for the local transaction.

Other backend won't be able to process any query against the impacted
relations until the transaction is finished as they will be blocked by the
AccessExclusiveLock acquired by this UTILITY, so there's no risk to create
invalid / inefficient plans from other backends.

FIXME: this is far from being finished.  Minimal list of things that still
needs to be handled:

 - way more regression tests, including isolationtester tests to test
   concurrent query execution behavior
 - find all the impacted relations in case of CASCADEd utility
 - this approach is only safe if the UTILITY requires an AccessExclusiveLock.
   Therefore any of the CONCURRENTLY version of UTILITY will void current
   assumption.  Find a way to deal with it :)
 - if this UTILITY is dropping one of the rtable entry, we should evict
   entirely the entry from the main hash table rather than discarding the
   plan.
 - we should also discard plans when new indexes are added
 - lot of work needed for partitioning
 - only RELATIONs are handled.  What about functions, operators...
 - client should not be able to override pg_shared_plans.read_only if we set
   it here.