Workaround for ALTER TEXT SEARCH DICTIONARY.
We have no way to track dependencies on TEXT SEARCH DICTIONNARY, so the only
way to avoid returning false result is to entirely drop the cache for the
current database.  Hopefully such commands should not be frequent so it won't
be a big issue.

Note that to properly test this behavior you need multiple backends as the one
executing the ALTER TEXT SEARCH DICTIONNARY wilL discard his own cache.

Note also that to ensure a correct behavior we have to make sure that the
command isn't run in a transaction (done here), and we will only drop the
cached plans after the UTILITY has been executed.

This implementation is not perect as it's still possible that someone locally
deactivate pg_shared_plans before executing such a command, which would be
bad, or that another backend cache a new plan between the reset and the commit
of the ALTER TEXT SELECT DICTIONARY.