Fix handling of reverse dependencies when a new plans is cached.
The previous code was assuming that when a new plan was cached, the reverse
dependencies would be the same, which is evidently wrong as a change in the
reverse dependency is the cause of a plan eviction.

This resulted in both incorrect reverse dependency tracking and dynamic shared
memory leak.