Switch back pg_shared_plans() to tuplestore mode.
The SRF was previously changed to use ValuePerCall mode to avoid using too many
resources, as it can result a big dataset.  But the design was rather fragile
as pgsp->lock could sometimes be released at a later point using the
ExprContext callback.  This is especially problematic using nesting
pg_shared_plans() with other SRFs that try to acquire pgsp->lock, as it can end
up with an infifinite deadlock on pgsp->lock.